+++ datalab/gui/actionhandler.py
@@ -646,53 +646,6 @@ class BaseActionHandler(metaclass=abc.ABCMeta):
                 select_condition=SelectCond.exactly_one_with_roi,
                 tip=_("Edit regions of interest numerically"),
             )
-            self.new_action(
-                _("Extract") + "...",
-                triggered=self.panel.processor.compute_roi_extraction,
-                # Icon name is 'signal_roi.svg' or 'image_roi.svg':
-                icon_name=f"{self.OBJECT_STR}_roi.svg",
-                separator=True,
-            )
-            self.new_action(
-                _("Copy"),
-                separator=True,
-                icon_name="roi_copy.svg",
-                tip=_("Copy regions of interest from selected %s") % self.OBJECT_STR,
-                triggered=self.panel.copy_roi,
-                select_condition=SelectCond.exactly_one_with_roi,
-                toolbar_pos=-1,
-            )
-            self.new_action(
-                _("Paste"),
-                icon_name="roi_paste.svg",
-                tip=_("Paste regions of interest into selected %s") % self.OBJECT_STR,
-                triggered=self.panel.paste_roi,
-                toolbar_pos=-1,
-            )
-            self.new_action(
-                _("Import") + "...",
-                icon_name="roi_import.svg",
-                tip=_("Import regions of interest into %s") % self.OBJECT_STR,
-                triggered=self.panel.import_roi_from_file,
-                select_condition=SelectCond.exactly_one,
-                toolbar_pos=-1,
-            )
-            self.new_action(
-                _("Export") + "...",
-                icon_name="roi_export.svg",
-                tip=_("Export selected %s regions of interest") % self.OBJECT_STR,
-                triggered=self.panel.export_roi_to_file,
-                select_condition=SelectCond.exactly_one_with_roi,
-                toolbar_pos=-1,
-            )
-            self.new_action(
-                _("Remove all"),
-                separator=True,
-                triggered=self.panel.processor.delete_regions_of_interest,
-                icon_name="roi_delete.svg",
-                select_condition=SelectCond.with_roi,
-                context_menu_pos=-1,
-            )
 
         with self.new_category(ActionCategory.VIEW):
             self.new_action(
@@ -786,6 +739,55 @@ class BaseActionHandler(metaclass=abc.ABCMeta):
 
     def create_last_actions(self):
         """Create actions that are added to the menus in the end"""
+        with self.new_category(ActionCategory.ROI):
+            self.new_action(
+                _("Extract") + "...",
+                triggered=self.panel.processor.compute_roi_extraction,
+                # Icon name is 'signal_roi.svg' or 'image_roi.svg':
+                icon_name=f"{self.OBJECT_STR}_roi.svg",
+                separator=True,
+            )
+            self.new_action(
+                _("Copy"),
+                separator=True,
+                icon_name="roi_copy.svg",
+                tip=_("Copy regions of interest from selected %s") % self.OBJECT_STR,
+                triggered=self.panel.copy_roi,
+                select_condition=SelectCond.exactly_one_with_roi,
+                toolbar_pos=-1,
+            )
+            self.new_action(
+                _("Paste"),
+                icon_name="roi_paste.svg",
+                tip=_("Paste regions of interest into selected %s") % self.OBJECT_STR,
+                triggered=self.panel.paste_roi,
+                toolbar_pos=-1,
+            )
+            self.new_action(
+                _("Import") + "...",
+                icon_name="roi_import.svg",
+                tip=_("Import regions of interest into %s") % self.OBJECT_STR,
+                triggered=self.panel.import_roi_from_file,
+                select_condition=SelectCond.exactly_one,
+                toolbar_pos=-1,
+            )
+            self.new_action(
+                _("Export") + "...",
+                icon_name="roi_export.svg",
+                tip=_("Export selected %s regions of interest") % self.OBJECT_STR,
+                triggered=self.panel.export_roi_to_file,
+                select_condition=SelectCond.exactly_one_with_roi,
+                toolbar_pos=-1,
+            )
+            self.new_action(
+                _("Remove all"),
+                separator=True,
+                triggered=self.panel.processor.delete_regions_of_interest,
+                icon_name="roi_delete.svg",
+                select_condition=SelectCond.with_roi,
+                context_menu_pos=-1,
+            )
+
         with self.new_category(ActionCategory.ANALYSIS):
             self.new_action(
                 _("Show results") + "...",
@@ -1007,6 +1009,14 @@ class ImageActionHandler(BaseActionHandler):
     def create_first_actions(self):
         """Create actions that are added to the menus in the first place"""
         super().create_first_actions()
+
+        with self.new_category(ActionCategory.ROI):
+            self.new_action(
+                _("Create ROI grid") + "...",
+                triggered=self.panel.processor.create_roi_grid,
+                icon_name="roi_grid.svg",
+                tip=_("Create a grid of regions of interest"),
+            )
 
         with self.new_category(ActionCategory.VIEW):
             showcontrast_action = self.new_action(